requirements/base.txt
Django~=2.2.12
psycopg2-binary==2.8.5
djangorestframework==3.11.0
djangorestframework-xml==1.3.0
django-debug-toolbar~=1.9.1
pyjwt==1.7.1
djangorestframework-simplejwt==4.4.0
django-model-utils==1.1.0
django-filter==2.4.0
pyyaml==5.2
django-extensions==3.1.1
mysql-connector==2.2.9
mysqlclient==2.0.3
drf-yasg==1.20.0
